% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_fastas.r
\name{blast_fastas}
\alias{blast_fastas}
\title{Run run_blastn}
\usage{
blast_fastas(x, sample_size = 1000)
}
\arguments{
\item{x}{a vector of fastas}

\item{sample_size}{the number of fastas to blast at once}
}
\value{
a data.frame representing the output of blastn
}
\description{
We're gonna leave it up to the user whether they want to randomize the input
}
