% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiedPrimerTreeFunctions.R
\name{BLAST_primer}
\alias{BLAST_primer}
\title{Query a pair of primers using ncbi's Primer-BLAST}
\usage{
BLAST_primer(
  forward,
  reverse,
  ...,
  organism = "",
  primer_specificity_database = "nt",
  exclude_env = "on"
)
}
\arguments{
\item{forward}{forward sequence}

\item{reverse}{reverse sequence}

\item{...}{passed onto the POST request (see NCBI for details)}

\item{organism}{organism}

\item{primer_specificity_database}{primer_specificity_database='nt'}

\item{exclude_env}{exclude_env='on'}
}
\value{
httr response object of the query, pass to parse_primer_hits to parse the results.
}
\description{
Query a pair of primers using ncbi's Primer-BLAST at https://www.ncbi.nlm.nih.gov/tools/primer-blast/
}
