% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_seeds.R
\name{blast_seeds}
\alias{blast_seeds}
\title{Wrapper function for blast_datatable that reads a data.frame from a path}
\usage{
blast_seeds(
  seeds_output_path,
  blast_db_path,
  accession_taxa_sql_path,
  working_dir,
  metabarcode_name,
  expand_vectors = TRUE,
  warnings = 0,
  ...
)
}
\arguments{
\item{seeds_output_path}{a path to a csv from get_seeds_local or get_seeds_remote}

\item{blast_db_path}{a directory containing a blast-formatted database}

\item{accession_taxa_sql_path}{a path to the accessionTaxa sql created by
taxonomizr}

\item{working_dir}{a directory in which to save partial and complete output}

\item{metabarcode_name}{a prefix for the fasta}

\item{expand_vectors}{logical, determines whether to expand too_many_Ns
and not_in db into real tables and write them in the output directory}

\item{warnings}{value to set the "warn" option to during the function call.
On exit it returns to the previous value. Setting this argument to
NULL will not change the option.}

\item{...}{additional arguments passed to \code{\link[=blast_datatable]{blast_datatable()}}}
}
\description{
blast_seeds uses \code{\link[=blast_datatable]{blast_datatable()}} to search against a blast
formatted database. It creates a directory at \code{working_dir} if one does not
already exist and generates \code{blast_seeds_output} and \code{.blast_seeds_save}
inside that directory. It passes \verb{<working_dir>/.blast_seeds_save} to
\code{\link[=blast_datatable]{blast_datatable()}} as the save directory and generates files in
\code{rblast_seeds_output} recording the results of the blast.
}
\section{Saving data}{
The \code{\link[=blast_datatable]{blast_datatable()}} call saves intermediate results and
metadata about the search as local files in the save directory generated by
blast_seeds. This allows the function to resume a partially
completed blast, mitigating the consequences of encountering an
error or experiencing other interruptions. To resume a partially completed
blast, supply the same seeds and working directory. See the documentation
of \code{\link[=blast_datatable]{blast_datatable()}} for more information.
}

