% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_seeds_multi_db.R
\name{blast_seeds_multi_db}
\alias{blast_seeds_multi_db}
\title{Run \code{blast_seeds} on each individual BLAST DB, potentially in-parallel.}
\usage{
blast_seeds_multi_db(
  blast_db_paths,
  output_directory_path,
  metabarcode_name,
  parallel_jobs,
  ...
)
}
\description{
Run \code{blast_seeds} on each individual BLAST DB, potentially in-parallel.
}
