% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_blast_plus_installation.R
\name{check_blast_plus_installation}
\alias{check_blast_plus_installation}
\title{Check BLAST+ installation}
\usage{
check_blast_plus_installation(ncbi_bin = NULL, return_success_message = FALSE)
}
\arguments{
\item{ncbi_bin}{path to NCBI binaries if they are not in the PATH
e.g. \verb{/my/local/ncbi-blast-2.10.1+/bin}. Leave blank or NULL otherwise}

\item{return_success_message}{whether to return a message when all binaries are found}
}
\value{
Nothing, a message or an error, depending on the outcome
}
\description{
Checks required BLAST+ binaries are in the PATH, or found with given path
}
