% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_fasta_and_taxonomy.R
\name{combine_fasta_and_taxonomy}
\alias{combine_fasta_and_taxonomy}
\title{Combines matching fasta and taxonomy files, calculates product length, and if need adds taxids}
\usage{
combine_fasta_and_taxonomy(
  output_directory_path,
  metabarcode_name,
  fasta_path,
  taxonomy_path,
  accession_taxa_sql_path,
  NCBI.accession = TRUE,
  add.taxid = TRUE
)
}
\arguments{
\item{output_directory_path}{the path to the output directory}

\item{metabarcode_name}{used to name the the files.}

\item{fasta_path}{the path to the fasta file of interest}

\item{taxonomy_path}{the path to the taxonomy file of interest}

\item{accession_taxa_sql_path}{the path to sql created by taxonomizr
(e.g. accession_taxa_sql_path <- "/my/accessionTaxa.sql")}

\item{NCBI.accession}{is name of the sequence an the NCBI accession number.
The default value is NCBI.accession = TRUE.  If NCBI.accession = FALSE
the function will asign taxonomy using species name.}

\item{add.taxid}{does the user want to add taxid. The default is add.taxid = TRUE, which adds
taxid bases on the status of NCBI.accession.}
}
\description{
combine_fasta_and_taxonomy takes a fasta file and a corresponding taxonomy file, and turns it
into an output .csv that can be used in \code{\link[=derep_and_clean_db]{derep_and_clean_db()}} and \code{rCRUX::compare_db()}.
}
\details{
Fasta files are converted to a dataframe with columns accession and sequence. The product length
for each seqeunce is calculated. Taxonomy is joined to the corresponding accession, and taxonomic
ranks are separated by superkingdom, phylum, class, order, family, and genus and species. Lastly
taxonomy is added based on NCBI accession, or species id if the accession is not available.

An rCRUX formatted summary .csv with the suffix "rCRUX_formatted_summary.csv" is generated in the
output directory indicated by output_directory_path.
}
\examples{

\dontrun{

output_directory_path <- "/my/directory/12S_mifish"
metabarcode_name <- "12S_mifish"
fasta_path = "/my/directory/12S_fasta_and_taxonomy/12S_.fasta"
taxonomy_path = "/my/directory/12S_fasta_and_taxonomy/12S_fasta_and_taxonomy/12S_taxonomy.txt"

combine_fasta_and_taxonomy(
 output_directory_path = output_directory_path
 metabarcode_name = metabarcode_name,
 fasta_path = fasta_path,
 taxonomy_path = taxonomy_path,
 accession_taxa_sql_path = accession_taxa_sql_path
)
}

}
