% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_ref_db.R
\name{compare_ref_db}
\alias{compare_ref_db}
\title{compare_ref_db() is a Reference database comparison function}
\usage{
compare_ref_db(
  ref_db_1_path,
  ref_db_2_path,
  ref_db_1_name,
  ref_db_2_name,
  out_dir,
  format_ref_db_1 = FALSE,
  format_ref_db_2 = FALSE
)
}
\arguments{
\item{ref_db_1_name}{a text string of ref_db_1 name}

\item{ref_db_2_name}{a text string of ref_db_2 name}

\item{out_dir}{a directory in which to create summary files and figures}

\item{format_ref_db_1}{indicates indicates if a user needs to format an output with separate
columns for accession, superkingdom, phylum, class, order, family, genus,
species, into a taxonomy.txt formatted file with two columns:
accession and superkingdom;phylum;class;order;family;genus;species
default value is FALSE (format_ref_db_2=FALSE)}

\item{format_ref_db_2}{indicates if a user needs to format an output with separate
columns for accession, superkingdom, phylum, class, order, family, genus,
species, into a taxonomy.txt formatted file with two columns:
accession and superkingdom;phylum;class;order;family;genus;species
default value is FALSE (format_ref_db_2=FALSE).}

\item{ref_db_1}{a data.frame formatted like the output from
blast_seeds() metabarcode_name_taxonomy.txt}

\item{ref_db_2}{a data.frame formatted like the output from
blast_seeds() metabarcode_name_taxonomy.txt}
}
\value{
\enumerate{
\item A data.frame representing the overlapping accessions of both reference databases
\item A data.frame representing the unique accessions in reference database 1
\item A data.frame representing the unique accessions in reference database 2
\item A venn diagram of accessions
\item Krona plots of the species found only in reference database 1 and 2
}
}
\description{
Compare two reference databases, produces a Venn Diagram,
Krona plot, and summary table of overlapping and mismatching accessions and associated taxonomy.

Note:
Krona plot requires the installation of krona tools \href{https://github.com/marbl/Krona/wiki/Installing}{krona tools} on your local computer.
}
\examples{

\dontrun{
ref_db_1_path <- "~/rCRUX_output/12S_parameters_test_1/get_seeds_local/12S_filtered_get_seeds_local_output_with_taxonomy.csv"
ref_db_2_path <- "~/rCRUX_output/12S_parameters_test_1/blast_seeds_output/12S_taxonomy.txt"
ref_db_1_name <- "get_seeds"
ref_db_2_name <- "blast_seeds"
out_dir <- "~/rCRUX_output/12S_parameters_test_1/comp_steps/"

compare_ref_db(ref_db_1_path,
              ref_db_2_path,
              ref_db_1_name,
              ref_db_2_name,
              out_dir, format_ref_db_1=TRUE)

ref_db_1_path <- "~/rCRUX_output/12S_parameters_test_1/blast_seeds_output/12S_taxonomy.txt"
ref_db_2_path <- "~/rCRUX_output/12S_parameters_test_2/blast_seeds_output/12S_taxonomy.txt"
ref_db_1_name <- "parameters_test_1"
ref_db_2_name <- "parameters_test_2"
out_dir <- "~/rCRUX_output/12S_parameters_tests/comp_parameters/"

compare_ref_db(ref_db_1_path,
              ref_db_2_path,
              ref_db_1_name,
              ref_db_2_name,
              out_dir)
}

}
