% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_qiime2.R
\name{convert_to_qiime2}
\alias{convert_to_qiime2}
\title{reformats the output of \code{\link[=blast_seeds]{blast_seeds()}}
and writes it to disk in a format compatible with Qiime2}
\usage{
convert_to_qiime2(
  ref_db_1_path,
  ref_db_1_name,
  out_dir,
  format_ref_db_1 = FALSE
)
}
\arguments{
\item{ref_db_1_path}{The path to a tab-separated file, formatted like the
output from \code{\link[=blast_seeds]{blast_seeds()}}}

\item{ref_db_1_name}{a text string of ref_db_1 name}

\item{out_dir}{a path to the directory where to write the output database}

\item{format_ref_db_1}{logical. If TRUE, assumes the input file comes in the
format of a comma-separated file with named columns for accession,
superkingdom, phylum, class, order, family, genus, species.
Defaults to FALSE (format_ref_db_1=FALSE)}
}
\description{
Converts an \code{\link[=blast_seeds]{blast_seeds()}} formatted database into a Qiime2
formatted database.

The functions reads the file into R, reformats it into Qiime desired format
and writes it in the desired path.
}
\details{
The input file will have two columns separated by a tab, and no headers. The
first column contains the accession of the sequences, the second column
includes the taxonomical annotation of those sequences, with the annotations
superkingom, phylum, class, order, family, genus and species in a single
string separated by colons.

The function reads in the input file using \code{\link[utils:read.table]{utils::read.table()}}, separates
the two fields and follows by separating the taxonomical fields into
columns.

With the option \code{format_ref_db_1} set to \code{TRUE} (the default), the function
expects a comma-separated file with separate columns for accession and the
following taxonomical ranks: superkingdom, phylum, class, order, family,
genus, species. The input database is reformatted using
\link{output_to_taxonomy_file}.

One the input file is in the desired format, the function reformats the
taxonomical fields by adding the prefixes that identified each taxonomical
rank.

The function returns no value into the R session, but writes to file on
disk.
}
\examples{
\dontrun{
convert_to_qiime2(
  ref_db_1_path = here(
    "eecurd_updated_dbs_20230420",
    "expanded_12S_20230420",
    "12S",
    "blast_seeds_output",
    "12S_taxonomy.txt"
  ),
  ref_db_1_name = "12S",
  out_dir = here()
)
}

}
