% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{create_sub_output_directory}
\alias{create_sub_output_directory}
\title{Create a subdirectory for multi DB pipeline output.}
\usage{
create_sub_output_directory(path, n)
}
\arguments{
\item{path}{The path where the subdirectory should be created}

\item{n}{The index of the DB split}
}
\value{
The path of the (possibly) created subdirectory
}
\description{
This function creates a named subdirectory within the specified path for
storing output files.  If the output directory already exists, no action is
taken, and the name of the directory is still returned.  (I.e., it is safe to
call on a directory that already exists.)
}
\examples{
create_sub_output_directory("/home/user/rCRUX_output", 1)
create_sub_output_directory("/home/user/rCRUX_output", 2)
create_sub_output_directory("/home/user/rCRUX_output", 3)
}
