% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dot-data.R
\name{.data}
\alias{.data}
\title{.data pronoun operator}
\value{
The result of calling \verb{function(column)} but using \verb{function(.data$column)}
}
\description{
See \link[rlang:dot-data]{rlang::.data} for details.
}
\keyword{internal}
