% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_multi_taxids.R
\name{expand_multi_taxids}
\alias{expand_multi_taxids}
\title{Expands returns from the nt database with multiple taxids}
\usage{
expand_multi_taxids(output_table, max_to_blast, blast_db_path, ncbi_bin = NULL)
}
\arguments{
\item{output_table}{table to use}

\item{max_to_blast}{passed to \code{\link[=blast_datatable]{blast_datatable()}} and is the maximum
number of entries to pass to blastcmdb.
The default is max_to_blast = 1000 - the optimal number of reads to
blast will depend on the user's environment (available RAM) and the
number of possible hits (determined by marker and parameters)}

\item{blast_db_path}{a directory containing one or more blast-formatted database.
For multiple blast databases, separate them with a space and add an extra set of quotes.
(e.g blast_db_path <- "/my/ncbi_nt/nt" or blast_db_path <- '"/my/ncbi_nt/nt  /my/ncbi_ref_euk_rep_genomes/ref_euk_rep_genomes"')}

\item{ncbi_bin}{the directory that the blast+ suite is in. If NULL, the
program will use your PATH environmental variable to locate them}
}
\value{
an output table
}
\description{
The blast db downloaded from NCBIs FTP site has representative accessions
meaning identical sequences have been collapsed across multiple accessions
even if they have different taxid.
Here we identify representative accessions with multiple taxids, and unpack
all of the accessions that go into that representitive accessions.
Note - we are not identifying or unpacking the representative accessions
that report a single taxid

The input table is similar to the summary.csv table produced by blast_seeds, but without taxonomy.
The output_table is in the same formate, but without multiple taxids in the blast taxids field
}
\examples{

\dontrun{
expand_multi_taxids(output_table, max_to_blast)
}
}
