% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seeds_local_multi_db.R
\name{get_seeds_local_multi_db}
\alias{get_seeds_local_multi_db}
\title{Run \code{get_seeds_local} on each individual BLAST DB, potentially in-parallel.}
\usage{
get_seeds_local_multi_db(
  blast_db_paths,
  output_directory_path,
  parallel_jobs,
  ...
)
}
\description{
Run \code{get_seeds_local} on each individual BLAST DB, potentially in-parallel.
}
