% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxonomizr_from_accession.R
\name{get_taxonomizr_from_accession}
\alias{get_taxonomizr_from_accession}
\title{Attach taxonomy data to an input table}
\usage{
get_taxonomizr_from_accession(input, accession_taxa_sql_path, organize = TRUE)
}
\arguments{
\item{input}{a data.frame}

\item{accession_taxa_sql_path}{the path to an accessionTaxa sql}
}
\value{
the data.frame with taxonomy data
}
\description{
This function will append taxids and taxonomy to any dataframe with a column
named accession that contains NCBI accessions (technically accession versions).
}
\details{
It takes a dataframe and searches for column called accession.  If present,
the column data is passed to \code{\link[taxonomizr:accessionToTaxa]{taxonomizr::accessionToTaxa()}} to find the
corresponding taxids. The taxids are passed to \code{\link[taxonomizr:getTaxonomy]{taxonomizr::getTaxonomy()}}
and the following taxonomic information is retrieved for each of the following
ranks: species, superkingdom, kingdom, phylum, subphylum, superclass, class,
subclass, order, family, subfamily, genus, infraorder, subcohort, superorder,
superfamily, tribe, subspecies, subgenus, species group, parvorder, varietas.
New columns for taxid and each rank are appended to the dataframe.
}
