% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxonomy_from_accession.R,
%   R/get_taxonomy_from_accession_unique.R
\name{get_taxonomy_from_accession}
\alias{get_taxonomy_from_accession}
\title{Attach taxonomy data to an input table using accession id}
\usage{
get_taxonomy_from_accession(
  input,
  accession_taxa_sql_path,
  arrange_taxonomy = TRUE
)

get_taxonomy_from_accession(
  input,
  accession_taxa_sql_path,
  arrange_taxonomy = TRUE
)
}
\arguments{
\item{input}{a data.frame with column \code{accession}}

\item{accession_taxa_sql_path}{the path to local SQL database created by \code{taxonomizr}}

\item{arrange_taxonomy}{arrange rows of the output by taxonomy (superkingdom -> species)}
}
\value{
the data.frame with taxonomy data

the data.frame with taxonomy data
}
\description{
This function will append taxids and taxonomy to any dataframe with a column
named accession that contains NCBI accessions (technically accession versions).

This function will append taxids and taxonomy to any dataframe with a column
named accession that contains NCBI accessions (technically accession versions).
}
\details{
It takes a dataframe and searches for a column called accession.  If present,
the column data is passed to \code{\link[taxonomizr:accessionToTaxa]{taxonomizr::accessionToTaxa()}} to find the
corresponding taxids. The taxids are passed to \code{\link[taxonomizr:getTaxonomy]{taxonomizr::getTaxonomy()}}
and the following taxonomic information is retrieved for each of the following
ranks: species, superkingdom, kingdom, phylum, subphylum, superclass, class,
subclass, order, family, subfamily, genus, infraorder, subcohort, superorder,
superfamily, tribe, subspecies, subgenus, species group, parvorder, varietas.
New columns for taxid and each rank are appended to the dataframe.

It takes a dataframe and searches for a column called accession.  If present,
the column data is passed to \code{\link[taxonomizr:accessionToTaxa]{taxonomizr::accessionToTaxa()}} to find the
corresponding taxids. The taxids are passed to \code{\link[taxonomizr:getTaxonomy]{taxonomizr::getTaxonomy()}}
and the following taxonomic information is retrieved for each of the following
ranks: species, superkingdom, kingdom, phylum, subphylum, superclass, class,
subclass, order, family, subfamily, genus, infraorder, subcohort, superorder,
superfamily, tribe, subspecies, subgenus, species group, parvorder, varietas.
New columns for taxid and each rank are appended to the dataframe.
}
