% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_blast_plus_binaries.R
\name{has_blast_plus_binaries}
\alias{has_blast_plus_binaries}
\title{Check BLAST+ installation}
\usage{
has_blast_plus_binaries(ncbi_bin = NULL)
}
\arguments{
\item{ncbi_bin}{path to NCBI binaries if they are not in the PATH}
}
\value{
TRUE/FALSE with attributes ncbi_bin, checked_binaries
}
\description{
Checks required BLAST+ binaries are in the PATH, or found with given path. This
function returns a TRUE/FALSE for quick logic, and is otherwise paired with
\code{\link[=check_blast_plus_installation]{check_blast_plus_installation()}} for detailed messaging.
}
