% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative_primer_search.R
\name{iterative_primer_search}
\alias{iterative_primer_search}
\title{Call primer_search with several parameters and aggregate the results}
\usage{
iterative_primer_search(
  forward_primer_seq,
  reverse_primer_seq,
  organisms,
  db = "nt",
  ...
)
}
\arguments{
\item{forward_primer_seq}{the forward primer to search (e.g. forward_primer_seq <- "TAGAACAGGCTCCTCTAG")}

\item{reverse_primer_seq}{the reverse primer to search (e.g. reverse_primer_seq <-  "TTAGATACCCCACTATGC")}

\item{organisms}{a character vector containing an id or name parseable by
NCBI as an organism. If it is a vector with multiple entries, each
entry will be queried separately.
(e.g. organism = c("1476529", "7776"))}

\item{db}{which NCBI database to search. The default is db = 'nt' (e.g. nr).}

\item{...}{arguments passed on to \code{\link[=primer_search]{primer_search()}}}
}
\value{
a data.table summarizing the results of several primer_searches
}
\description{
This function acts like \code{primerTree::primer_search()} and
primerTree's parse_primer() hits all in one.
}
\details{
It calls \code{rCRUX::modifiedPrimerTree_Functions()} to perform tasks. Its parameters
are very similar to \code{primerTree::primer_search()}, but it takes vectors for
organism and for database and performs a primer search for each combination.
It downgrades errors from primer_search and parse_primer_hits into warnings.
This is useful when searching for a large number of different combinations,
allowing the function to output successful results.
}
