% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_to_taxonomy_file.R
\name{output_to_taxonomy_file}
\alias{output_to_taxonomy_file}
\title{output_to_taxonomy_file
function to turn a table with separate accession, s, p, o, c, f, g, s columns
into rCRUX taxonomy file with two columns: accession and s;p;c;o;f;g;s}
\usage{
output_to_taxonomy_file(table_path, metabarcode, out_dir)
}
\arguments{
\item{table_path}{the path}

\item{metabarcode}{the code}

\item{out_dir}{the output path}
}
\value{
nothing
}
\description{
output_to_taxonomy_file
function to turn a table with separate accession, s, p, o, c, f, g, s columns
into rCRUX taxonomy file with two columns: accession and s;p;c;o;f;g;s
}
