% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiedPrimerTreeFunctions.R
\name{primer_search}
\alias{primer_search}
\title{Query a pair of primers using NCBI's Primer-BLAST, if primers contain iupac}
\usage{
primer_search(
  forward,
  reverse,
  num_permutations = 25,
  ...,
  .parallel = FALSE,
  .progress = "none"
)
}
\arguments{
\item{forward}{forward primer to search by 5'-3' on plus strand}

\item{reverse}{reverse primer to search by 5'-3' on minus strand}

\item{num_permutations}{the number of primer permutations to search, if the degenerate bases
cause more than this number of permutations to exist, this number will be
sampled from all possible permutations.}

\item{...}{additional arguments passed to Primer-Blast, BLAST_primer}

\item{.parallel}{if 'TRUE', perform in parallel, using parallel backend
provided by foreach}

\item{.progress}{name of the progress bar to use, see 'create_progress_bar'}
}
\value{
a list of results for each primer?
}
\description{
Query a pair of primers using NCBI's Primer-BLAST, if primers contain iupac
}
\details{
PrimerTree
Modified 2022-08-10 by Luna Gal
Removed num_aligns argument from primer_search
Instead num_targets_with_primers is determined by the user
Changed the calculation for duration of blast submission

ambiguity codes, enumerate all possible combinations and combine the
results.
}
