% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrux_blast.R
\name{rcrux_blast}
\alias{rcrux_blast}
\title{Wrapper function for blast_datatable that reads a data.frame from a path}
\usage{
rcrux_blast(
  seeds_path,
  db_dir,
  accession_taxa_path,
  working_dir,
  metabarcode,
  expand_vectors = TRUE,
  warnings = 0,
  ...
)
}
\arguments{
\item{seeds_path}{a path to a csv from get_blast_seeds}

\item{db_dir}{a directory containing a blast-formatted database}

\item{accession_taxa_path}{a path to the accessionTaxa sql created by
taxonomizr}

\item{working_dir}{a directory in which to save partial and complete output}

\item{metabarcode}{a prefix for the fasta}

\item{expand_vectors}{logical, determines whether to expand too_many_Ns
and not_in db into real tables and write them in the output directory}

\item{warnings}{value to set the "warn" option to during the function call.
On exit it returns to the previous value. Setting this argument to
NULL will not change the option.}

\item{...}{additional arguments passed to \code{\link[RCRUX.dev:blast_datatable]{RCRUX.dev::blast_datatable()}}}
}
\description{
rcrux_ blast uses \code{\link[RCRUX.dev:blast_datatable]{RCRUX.dev::blast_datatable()}} to search against a blast
formatted database. It creates a directory at \code{working_dir} if one does not
already exist and generates \code{rcrux_blast_output} and \code{.rcrux_blast_save}
inside that directory. It passes \verb{<working_dir>/.rcrux_blast_save} to
\code{\link[RCRUX.dev:blast_datatable]{RCRUX.dev::blast_datatable()}} as the save directory and generates files in
\code{rcrux_blast_output} recording the results of the blast.
}
\section{Saving data}{
The \code{\link[RCRUX.dev:blast_datatable]{RCRUX.dev::blast_datatable()}} call saves intermediate results and
metadata about the search as local files in the save directory generated by
rcrux_blast. This allows the function to resume a partially
completed blast, mitigating the consequences of encountering an
error or experiencing other interruptions. To resume a partially completed
blast, supply the same seeds and working directory. See the documentation
of \code{\link[RCRUX.dev:blast_datatable]{RCRUX.dev::blast_datatable()}} for more information.
}

