% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representative_fasta_and_taxonomy.R
\name{representative_fasta_and_taxonomy}
\alias{representative_fasta_and_taxonomy}
\title{Get representative_fasta_and_taxonomy}
\usage{
representative_fasta_and_taxonomy(
  paths_to_summary_tables,
  metabarcode_name,
  output_directory_path
)
}
\arguments{
\item{paths_to_summary_tables}{character vector with 2 values}

\item{metabarcode_name}{name}

\item{output_directory_path}{output path}
}
\description{
function to get the representative fasta and taxonomy files for derep_and_clean_db
}
