% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_basic_pipeline.R
\name{run_basic_pipeline}
\alias{run_basic_pipeline}
\title{Run the standard rCRUX pipeline.}
\usage{
run_basic_pipeline(
  forward_primer_seq,
  reverse_primer_seq,
  metabarcode_name,
  output_directory_path,
  accession_taxa_sql_path,
  blast_db_path,
  args_get_seeds_local,
  args_blast_seeds
)
}
\description{
Run the standard rCRUX pipeline.
}
