% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_blastdbcmd.R
\name{run_blastdbcmd}
\alias{run_blastdbcmd}
\title{Extracts seed amplicon sequence from blast databases using blastdbcmd and
primer start and end coordinates stored in the seeds table}
\usage{
run_blastdbcmd(query_row, db, ncbi_bin = NULL)
}
\arguments{
\item{query_row}{a row from get_seeds_local or get_seeds_remote}

\item{db}{a directory with a blast-formatted database}

\item{ncbi_bin}{is the path to blast+ tools if not in the user's path.
Specify only if blastn and blastdbcmd  are not in your path.
The default is ncbi_bin = NULL - if not specified in path do the
following: ncbi_bin = "/my/local/blast+_folder".}
}
\value{
a fasta-formatted character vector
}
\description{
Given a row from a blast_seed formatted data.frame, extracts arguments
for blastdbcmd, then returns the output of the blastdbcmd call.
}
