% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_blastn.R
\name{run_blastn}
\alias{run_blastn}
\title{Run blastn with the seed amplicon sequences}
\usage{
run_blastn(
  fasta,
  db,
  temp_fasta_path = NULL,
  ncbi_bin = NULL,
  evalue = 1e-06,
  align = 50000,
  coverage = 50,
  perID = 70,
  num_threads = 1
)
}
\arguments{
\item{fasta}{a fasta-formatted string}

\item{db}{a path to a directory / directories containing one or more blast-formatted database.
For multiple blast databases, separate them with a space and add an extra set of quotes.
(e.g blast_db_path <- "/my/ncbi_nt/nt" or blast_db_path <- '"/my/ncbi_nt/nt  /my/ncbi_ref_euk_rep_genomes/ref_euk_rep_genomes"')}

\item{temp_fasta_path}{a file path to write a temporary fasta to - blastn requires a file to query}

\item{ncbi_bin}{is the path to blast+ tools if not in the user's path.
Specify only if blastn and blastdbcmd  are not in your path.
The default is ncbi_bin = NULL - if not specified in path do the
following: ncbi_bin = "/my/local/ncbi-blast-2.10.1+/bin/".}

\item{evalue}{is the number of expected hits with a similar quality score
found by chance. The default is evalue = 1e-6.}

\item{align}{is the maximum number of subject hits to return per query
blasted. The default is align = 50000.}

\item{coverage}{is the minimum percent of the query length recovered in the
subject hits. The default is coverage = 50.}

\item{perID}{is the minimum percent identity of the query relative to the
subject hits. The default is perID = 70.}

\item{num_threads}{number, the number of CPUs to engage in the blastn search. The
value 'max' can be used and which uses \code{\link[parallel:detectCores]{parallel::detectCores()}} to determine
the user's maximum number of CPUs automatically (use with caution; Default = 1)}
}
\value{
a tibble representing the blastn results
}
\description{
Calls \code{blastn} with a fasta file created from a character vector as the query.
The user can not add additional search parameters, but can modify the available parameters.
}
\details{
\code{\link[=run_blastn]{run_blastn()}} takes a character vector fasta, converts to a  fasta file,
and queries them against a BLAST formatted database. The result is an output table with
the following columns of data: accession, amplicon_length, pident,
query_accession, accession_sequence_length, amplicon_start, amplicon_stop,
sequence, evalue, BLAST_db_taxids.

Information about the blastn parameters can be found by accessing blastn -help
and at \href{https://www.ncbi.nlm.nih.gov/books/NBK279684/}{NCBI}.
}
