% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_multi_db_pipeline.R
\name{run_multi_db_pipeline}
\alias{run_multi_db_pipeline}
\title{Run the rCRUX multi-database pipeline.}
\usage{
run_multi_db_pipeline(
  forward_primer_seq,
  reverse_primer_seq,
  metabarcode_name,
  output_directory_path,
  accession_taxa_sql_path,
  blast_db_paths,
  parallel_jobs,
  args_get_seeds_local_multi_db,
  args_blast_seeds_multi_db
)
}
\description{
Run the rCRUX multi-database pipeline.
}
