% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_output_as_csv.R
\name{save_output_as_csv}
\alias{save_output_as_csv}
\title{Function to save data in .csv files}
\usage{
save_output_as_csv(file_name, description, file_out, Metabarcode)
}
\arguments{
\item{file_name}{the object to write}

\item{description}{a string to include in the .csv name
directly before the extension}

\item{file_out}{a parent directory; it should end in a /}

\item{Metabarcode}{a label that appears before the file name}
}
\value{
The value of write.table
}
\description{
Calls write.table with file_name as the first argument
and <file_out><Metabarcode><description>.csv as the second
}
