% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_sample.R
\name{smart_sample}
\alias{smart_sample}
\title{Wrapper function for sample
If the sample size is larger than the population, returns the input vector
Otherwise returns the result of sampling}
\usage{
smart_sample(x, size, ...)
}
\arguments{
\item{x}{a vector to sample from}

\item{size}{the size of the sample}

\item{...}{additional arguments to pass to sample}
}
\value{
a vector sampled from x
}
\description{
Wrapper function for sample
If the sample size is larger than the population, returns the input vector
Otherwise returns the result of sampling
}
