% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_blast_db.R
\name{check_blast_db}
\alias{check_blast_db}
\title{Check BLAST DB exists}
\usage{
check_blast_db(path)
}
\arguments{
\item{path}{path to BLAST DB}
}
\value{
Nothing or an error
}
\description{
The path of a BLAST DB should be the directory path + the name of the DB e.g
\code{path/to/blast-db} + \code{myDB} = \code{path/to/blast-db/myDB}. Usually the files in the
BLAST DB start with the name of the the DB e.g \code{myDB.ndb}, \code{myDB.ntf} ...
}
