% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_primer_hits.r
\name{filter_primer_hits}
\alias{filter_primer_hits}
\title{Remove empty accession numbers and other problematic rows}
\usage{
filter_primer_hits(
  hits_table,
  forward_primer,
  reverse_primer,
  mismatch = 3,
  minimum_length = 5,
  maximum_length = 500
)
}
\arguments{
\item{hits_table}{a data.frame generated by parse_primer_hits
or by iterative_primer_search}

\item{mismatch}{the maximum allowable value (inclusive) in the
mismatch_forward and mismatch_reverse columns}
}
\value{
a data.table with problematic rows removed
}
\description{
Remove empty accession numbers and other problematic rows
}
