% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seeds_local.R
\name{get_seeds_local}
\alias{get_seeds_local}
\title{Local blastn interpretation of querying primer_blast to and generate a .csv to use for blast_seeds}
\usage{
get_seeds_local(
  forward_primer_seq,
  reverse_primer_seq,
  output_directory_path,
  metabarcode_name,
  accession_taxa_sql_path,
  blast_db_path,
  mismatch = 6,
  minimum_length = 5,
  maximum_length = 500,
  primer_specificity_database = "nt",
  ...,
  return_table = TRUE
)
}
\arguments{
\item{forward_primer_seq}{passed to primer_to_fasta, which turns it into fasta
file to be past to get_seeds_local}

\item{reverse_primer_seq}{passed to primer_to_fasta, which turns it into fasta
file to be past to get_seeds_local}

\item{output_directory_path}{the parent directory to place the data in.}

\item{metabarcode_name}{used to name the subdirectory and the files. If a
directory named metabarcode_name does not exist in output_directory_path, a
new directory will be created. get_seeds_remote appends
metabarcode_name to the beginning of each of the two files it
generates.}

\item{accession_taxa_sql_path}{the path to sql created by taxonomizr}

\item{blast_db_path}{a directory with a blast-formatted database}

\item{mismatch}{the highest acceptable mismatch value per hit. get_seeds_local removes each
row with a mismatch greater than the specified value.}

\item{minimum_length}{get_seeds_local removes each row that has a value less than
minimum_length in the product_length column.}

\item{maximum_length}{get_seeds_local removes each row that has a
value greater than maximum_length in the product_length column}

\item{task}{the task for blastn to perform - default here is "blastn_short",
which is optimized for searches with queries < 50 bp}

\item{word_size}{is the fragment size used for blastn search - smaller word
sizes increase sensitivity and time of the search - default value is 7}

\item{evalue}{is the number of expected hits with a similar quality score
found by chance - default is 3e-7.}

\item{coverage}{is the minimum percent of the query length recovered in the
subject hits}

\item{perID}{is the minimum percent identity of the query relative to the
subject hits, the default is 50}

\item{reward}{is the reward for nucleotide match, the default is 2}
}
\value{
a data.frame containing the same information as the .csv it generates
}
\description{
get_seeds_local is a local interpretation of get_seeds_remote that avoids
querying NCBI's \href{https://www.ncbi.nlm.nih.gov/tools/primer-blast/}{primer BLAST}
tool. Although it is slower than remotly genertating blast seeds, it is not
subject to the arbitrary throttling of jobs that require significant memory.
It creates a directory at \code{output_directory_path} if one doesn't yet
exist, then creates a subdirectory inside \code{output_directory_path} named after
\code{metabarcode_name}. It creates two files inside that directory, one
representing the output and the other representing the output without added
taxonomy.

Information about the blastn parameters can be found in run_primer_blast, and
by accessing blastn -help.  Default parameters were optimized to provide
results similar or expanded results that through remote blast via primer-blast.
}
