% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seeds_remote.R
\name{get_seeds_remote}
\alias{get_seeds_remote}
\title{Query primer NCBI's Blast and generate a .csv to use for blast_seeds}
\usage{
get_seeds_remote(
  forward_primer_seq,
  reverse_primer_seq,
  output_directory_path,
  metabarcode_name,
  accession_taxa_sql_path,
  organism,
  mismatch = 3,
  minimum_length = 5,
  maximum_length = 500,
  primer_specificity_database = "nt",
  ...,
  return_table = TRUE
)
}
\arguments{
\item{forward_primer_seq}{passed to \code{rCRUX::primer_search()}, which turns it
into a list of all possible non degenerate primers, then passes
a user defined number of primer set combinations to NCBI.
(e.g. forward_primer_seq <- "TAGAACAGGCTCCTCTAG")}

\item{reverse_primer_seq}{passed to \code{rCRUX::primer_search()}, which turns it
into a list of all possible non degenerate primers, then passes
a user defined number of primer set combinations to NCBI.
(e.g. reverse_primer_seq <-  "TTAGATACCCCACTATGC")}

\item{output_directory_path}{the parent directory to place the data in.
(e.g. "/path/to/output/12S_V5F1_remote_111122")}

\item{metabarcode_name}{is passed to \code{\link[=get_seeds_remote]{get_seeds_remote()}} which appends
metabarcode_name to the beginning of each of the two files it
generates (e.g. metabarcode_name <- "12S_V5F1").}

\item{accession_taxa_sql_path}{the path to sql created by taxonomizr
(e.g. accession_taxa_sql_path <- "/my/accessionTaxa.sql")}

\item{organism}{a vector of character vectors. Each character vector is
passed in turn to primer_search, which passes them to NCBI.
get_seeds_remote aggregates all of the results into a single file.
(e.g. organism = c("1476529", "7776")) - note increasing taxonomic
rank (e.g. increasing from order to class) for this parameter can
maximize primer hits, but can also lead to API run throttling due to
memory limitations}

\item{mismatch}{the highest acceptable mismatch value. parse_primer_hits
returns a table with a mismatch column. get_seeds_remote removes each
row with a mismatch greater than the specified value.
The default is mismatch = 3 - Note this is smaller than \code{\link[=get_seeds_local]{get_seeds_local()}}}

\item{minimum_length}{\code{rCRUX::parse_primer_hits()} returns a table with a product_length
column. get_seeds_remote removes each row that has a value less than
minimum_length in the product_length column.
The default is minimum_length = 5}

\item{maximum_length}{\code{rCRUX::parse_primer_hits()} returns a table with a
product_length column. get_seeds_remote removes each row that has a
value greater than maximum_length in the product_length column
The default is maximum_length = 500}

\item{primer_specificity_database}{passed to \code{\link[=primer_search]{primer_search()}}, which passes it
to NCBI.  The default is primer_specificity_database = 'nt'.}

\item{...}{additional arguments passed to \code{\link[=iterative_primer_search]{iterative_primer_search()}}, see
\code{primerTree::primer_search()} and \href{https://www.ncbi.nlm.nih.gov/tools/primer-blast/}{NCBI primer-blast tool}
for more information.}

\item{return_table}{whether to return the table \verb{some table}}
}
\value{
a data.frame containing the same information as the .csv it generates
}
\description{
get_seeds_remote combines modified versions of \code{primerTree::primer_search()}
and primerTree's parse_primer to make \code{\link[=iterative_primer_search]{iterative_primer_search()}}
which is called to query NCBI's
\href{https://www.ncbi.nlm.nih.gov/tools/primer-blast/}{primer BLAST}
tool, filters the results, then aggregates them into a single data.frame.
It creates a directory \code{get_seeds_remote} in the \code{output_directory_path}.
It creates three files inside that directory. One represents the unfiltered
output and another represents the output after filtering with user modifiable
parameters and with appended taxonomy. Also generated is a summary of unique
taxonomic ranks after filtering.
}
\details{
get_seeds_remote passes the forward and reverse primer sequence for a given
PCR product to \code{\link[=iterative_primer_search]{iterative_primer_search()}} along with the taxid(s) of
the organism(s) to blast, the database to search, and many additional possible
parameters to NCBI's primer blast tool (see Note below). Degenerate primers
are converted into all possible non degenerate sets and a user defined maximum
number of primer combinations is passed to to the API. Multiple taxids are
searched independently, as are multiple database searches (e.g. nt and
refseq_representative_genomes). The data are parsed and stored in a dataframe,
which is also written to a file with the suffix
\verb{_unfiltered_get_seeds_remote_output.csv}.

These hits are further filtered using \code{\link[=filter_primer_hits]{filter_primer_hits()}} to
calculate and append amplicon size to the dataframe. Only hits that pass with default
or user modified length and number of mismatches parameters are retained.

Taxonomy is appended to these filtered hits using
\code{\link[=get_taxonomy_from_accession]{get_taxonomy_from_accession()}}. The results are written to
to file with the suffix \verb{_filtered_get_seeds_remote_output_with_taxonomy.csv}.
The number of unique instances for each rank in the taxonomic path for the
filtered hits are tallied (NAs are counted once per rank) and written to a
file with the suffix \verb{_filtered_get_seeds_local_remote_taxonomic_rank_counts.txt}

Note:
get_seeds_remote passes many parameters to NCBI's primer blast tool.
You can match the parameters to the fields available in the GUI
\href{https://www.ncbi.nlm.nih.gov/tools/primer-blast/}{here}. First, use your
browser to view the page source. Search for the field you are interested in
by searching for the title of the field. It should be enclosed in a tag.
Inside the label tag, it says \verb{for = "<name_of_parameter>"}. Copy the string
after for = and add it to get_seeds_remote as the name of a parameter, setting
it equal to whatever you like.

As of 2022-08-16, the primer blast GUI contains some options that are not
implemented by \code{primerTree::primer_search()} and by extension \code{\link[=iterative_primer_search]{iterative_primer_search()}}
primer_search doesn't include explicit documentation of allowed options, but
it will quickly report if an option isn't allowed, so trial and error will
not be very time consuming.

Note:
See \code{\link[=iterative_primer_search]{iterative_primer_search()}} and \code{rCRUX::modifiedPrimerTree_Functions}
for additional run parameters not included below.

Check NCBI's primer blast for additional search options**

get_seeds_remote passes many parameters to NCBI's primer blast tool. You can
match the parameters to the fields available in the GUI here. First, use your
browser to view the page source. Search for the field you are interested in
by searching for the title of the field. It should be enclosed in a tag.
Inside the label tag, it says for = "<name_of_parameter>". Copy the string
after for = and add it to get_seeds_remote as the name of a parameter,
setting it equal to whatever you like.

As of 2022-08-16, the primer blast GUI contains some options that are not implemented
by primer_search. The table below documents some of the available options.\tabular{ll}{
   Name \tab Default \cr
   PRIMER_SPECIFICITY_DATABASE \tab nt \cr
   EXCLUDE_ENV \tab unchecked \cr
   ORGANISM \tab Homo sapiens \cr
   TOTAL_PRIMER_SPECIFICITY_MISMATCH \tab 1 \cr
   PRIMER_3END_SPECIFICITY_MISMATCH \tab 1 \cr
   TOTAL_MISMATCH_IGNORE \tab 6 \cr
   MAX_TARGET_SIZE \tab 4000 \cr
   HITSIZE \tab 50000 \cr
   EVALUE \tab 30000 \cr
   WORD_SIZE \tab 7 \cr
   NUM_TARGETS_WITH_PRIMERS \tab 1000 \cr
   MAX_TARGET_PER_TEMPLATE \tab 100 \cr
}
}
\examples{
\dontrun{
forward_primer_seq = "TAGAACAGGCTCCTCTAG"
reverse_primer_seq =  "TTAGATACCCCACTATGC"
output_directory_path <- "/my/directory/12S_V5F1_remote_111122_modified_params"
metabarcode_name <- "12S_V5F1"
accession_taxa_sql_path <- "/my/directory/accessionTaxa.sql"


get_seeds_remote(forward_primer_seq,
                reverse_primer_seq,
                output_directory_path,
                metabarcode_name,
                accession_taxa_sql_path,
                HITSIZE ='1000000',
                evalue='100000',
                word_size='6',
                MAX_TARGET_PER_TEMPLATE = '5',
                NUM_TARGETS_WITH_PRIMERS ='500000', minimum_length = 50,
                MAX_TARGET_SIZE = 200,
                organism = c("1476529", "7776"), return_table = FALSE)


# This results in approximately 111500 blast seed returns (there is some 
# variation due to database updates, etc.), note the default generated 
# approximately 1047. This assumes the user is not throttled by memory 
# limitations.
}
}
