% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxonomizr_from_accession.R
\name{get_taxonomizr_from_accession}
\alias{get_taxonomizr_from_accession}
\title{Attach taxonomy data to an input table}
\usage{
get_taxonomizr_from_accession(input, accessionTaxa_path, organize = TRUE)
}
\arguments{
\item{input}{a data.frame}

\item{accessionTaxa_path}{the path to an accessionTaxa sql}
}
\value{
the data.frame with taxonomy data
}
\description{
Attach taxonomy data to an input table
}
