% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative_primer_search.r
\name{iterative_primer_search}
\alias{iterative_primer_search}
\title{Call primer_search with several parameters and aggregate the results}
\usage{
iterative_primer_search(
  forwar_primer_seq,
  reverse_primer_seq,
  organisms,
  db = "nt",
  ...
)
}
\arguments{
\item{reverse_primer_seq}{the reverse primer}

\item{organisms}{a character vector containing an id or name parseable by
NCBI as an organism. If it is a vector with multiple entries, each
entry will be queried separately.}

\item{db}{which NCBI database to search. The default is db - 'nt' (e.g. nr).}

\item{forward_primer_seq}{the forward primer}
}
\value{
a data.table summarizing the results of several primer_searches
}
\description{
This function acts like \code{\link[primerTree:primer_search]{primerTree::primer_search()}} and
primerTree's parse_primer() hits all in one. It calls
\code{\link[=modifiedPrimerTree_Functions]{modifiedPrimerTree_Functions()}} to perform tasks. Its parameters
are very similar to \code{\link[primerTree:primer_search]{primerTree::primer_search()}}, but it takes vectors for
organism and for database and performs a primer search for each combination.
It downgrades errors from primer_search and parse_primer_hits into warnings.
This is useful when searching for a large number of different combinations,
allowing the function to output successful results.
}
