% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative_primer_search.r
\name{iterative_primer_search}
\alias{iterative_primer_search}
\title{Call primer_search with several parameters and aggregate the results}
\usage{
iterative_primer_search(forward, reverse, organisms, db = "nt", ...)
}
\arguments{
\item{forward}{the forward primer}

\item{reverse}{the reverse primer}

\item{organisms}{a character vector containing an id or name parseable by
NCBI as an organism. If it is a vector with multiple entries, each
entry will be queried separately.}

\item{db}{which NCBI database to search.}
}
\value{
a data.table summarizing the results of several primer_searches
}
\description{
This function acts like primer_search and parse_primer hits all in one. Its
parameters are very similar to primer_search, but it takes vectors for
organism and for database and performs a primer search for each combination.
It downgrades errors from primer_search and parse_primer_hits into warnings.
This is useful when searching for a large number of different combinations,
allowing the function to output successful results.
}
