% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_blastdbcmd.R
\name{run_blastdbcmd}
\alias{run_blastdbcmd}
\title{Extracts query parameters from a row and calls blastdbcmd}
\usage{
run_blastdbcmd(query_row, db, ncbi_bin = NULL)
}
\arguments{
\item{query_row}{a row from get_blast_seeds}

\item{db}{a directory with a blast-formatted database}

\item{ncbi_bin}{if not null use it as the parent directory for blastn}
}
\value{
a fasta-formatted character vector
}
\description{
Given a row from a blast_seed formatted data.frame, extracts arguments
for blastdbcmd, then returns the output of the blastdbcmd call.
}
