% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_blastn.R
\name{run_blastn}
\alias{run_blastn}
\title{Runs blastn with the seed amplicon input fasta as a query}
\usage{
run_blastn(
  fasta,
  db_dir,
  temp = NULL,
  ncbi_bin = NULL,
  evalue = 1e-06,
  align = 50000,
  coverage = 50,
  perID = 70
)
}
\arguments{
\item{fasta}{a fasta-formatted string}

\item{ncbi_bin}{is the path to blast+ tools if not in the user's path.
Specify only if blastn and blastdbcmd  are not in your path.
The default is ncbi_bin = NULL - if not specified in path do the
following: ncbi_bin = "/my/local/blast+_folder".}

\item{evalue}{is the number of expected hits with a similar quality score
found by chance. The default is evalue = 1e-6.}

\item{align}{is the maximum number of subject hits to return per query
blasted. The default is align = 50000.}

\item{coverage}{is the minimum percent of the query length recovered in the
subject hits. The default is coverage = 50.}

\item{perID}{is the minimum percent identity of the query relative to the
subject hits. The default is perID = 70.}

\item{temp:}{a file path to write a temporary fasta to. The default is
temp = NULL.}
}
\value{
a tibble representing the blastn results
}
\description{
Runs blastn with the seed amplicon input fasta as a query
}
\details{
Calls blastn with a fasta file as the query. The user can not add
additional search parameters, but can modify the available parameters.

run_blastn takes a fasta file, and query them to a blast formatted database.
The result is an output table with
the following columns of data: accession, amplicon_length, pident,
query_accession, accession_sequence_length, amplicon_start, amplicon_stop,
sequence, evalue, BLAST_db_taxids.

Information about the blastn parameters can be found by accessing blastn -help
and at \href{https://www.ncbi.nlm.nih.gov/books/NBK279684/}{NCBI}.
}
