% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_blastn.R
\name{run_blastn}
\alias{run_blastn}
\title{Runs blastn with the input fasta as a query}
\usage{
run_blastn(
  fasta,
  db_dir,
  temp = "temp.fasta",
  ncbi_bin = NULL,
  evalue = 1e-06,
  align = 50000,
  coverage = 50,
  perID = 70
)
}
\arguments{
\item{fasta}{a fasta-formatted string}

\item{temp:}{a file path to write a temporary fasta to}

\item{ncbi_bin:}{if not null use it as the parent directory for blastn}
}
\value{
a tibble representing the blastn results
}
\description{
Takes the input string, writes it to a temporary file,
and calls blastn with that file as the query. Warning:
if a file at the path specified by temp already exists,
it will be overwritten then deleted.
}
