% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_primer_blastn.R
\name{run_primer_blastn}
\alias{run_primer_blastn}
\title{Runs blastn with the input primer sequences converted to a primer_fasta file
as a query}
\usage{
run_primer_blastn(
  primer_fasta,
  db,
  ncbi_bin = NULL,
  task = "blastn-short",
  word_size = 7,
  evalue = "3e+07",
  coverage = 90,
  perID = 50,
  reward = 2
)
}
\arguments{
\item{primer_fasta}{path to the primer fasta file}

\item{db}{path to blast formatted database}

\item{ncbi_bin}{if not null use it as the parent directory for blastn}

\item{task}{the task for blastn to perform - default here is "blastn_short",
which is optimized for searches with queries < 50 bp}

\item{word_size}{is the fragment size used for blastn search - smaller word
sizes increase sensitivity and time of the search.
The default is word_size =  7}

\item{evalue}{is the number of expected hits with a similar quality score
found by chance. The default is evalue = 3e-7.}

\item{coverage}{is the minimum percent of the query length recovered in the
subject hits. The default is coverage = 90.}

\item{perID}{is the minimum percent identity of the query relative to the
subject hits. The default is perID = 2.}

\item{reward}{is the reward for nucleotide match. The default is reward = 2.}
}
\value{
a tibble 'output_table' representing the blastn results
}
\description{
Takes the input string, writes it to a temporary file,
and calls blastn with that file as the query. Warning:
if a file at the path specified by temp already exists,
it will be overwritten then deleted.

The number of alignments returned for a given blast search is hardcoded at
"-num_alignments", "10000000",
}
