% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_primer_blastn.R
\name{run_primer_blastn}
\alias{run_primer_blastn}
\title{Runs blastn with the input primer sequences converted to a primer_fasta file
as a query}
\usage{
run_primer_blastn(
  primer_fasta,
  db,
  ncbi_bin = NULL,
  task = "blastn-short",
  word_size = 7,
  evalue = "3e+07",
  align = "10000000",
  coverage = 90,
  perID = 50,
  reward = 2,
  num_threads = NULL
)
}
\arguments{
\item{primer_fasta}{path to the primer fasta file}

\item{db}{a path to a directory / or directories containing one or more blast-formatted database.
For multiple blast databases, separate them with a space and add an extra set of quotes.
(e.g blast_db_path <- "/my/ncbi_nt/nt" or blast_db_path <- '"/my/ncbi_nt/nt  /my/ncbi_ref_euk_rep_genomes/ref_euk_rep_genomes"')}

\item{ncbi_bin}{is the path to blast+ tools if not in the user's path.
Specify only if blastn and blastdbcmd  are not in your path.
The default is ncbi_bin = NULL - if not specified in path do the
following: ncbi_bin = "/my/local/ncbi-blast-2.10.1+/bin/".}

\item{task}{the task for blastn to perform - default here is "blastn_short",
which is optimized for searches with queries < 50 bp}

\item{word_size}{is the fragment size used for blastn search - smaller word
sizes increase sensitivity and time of the search.
The default is word_size =  7}

\item{evalue}{is the number of expected hits with a similar quality score
found by chance. The default is evalue = '3e-7'.}

\item{align}{is the maximum number of subject hits to return per query
blasted. The default is align = '10000000'. - to few alignments will
result in no matching pairs of forward and reverse primers.  To many
alignments can result in an error due to RAM limitations.}

\item{coverage}{is the minimum percent of the query length recovered in the
subject hits. The default is coverage = 90.}

\item{perID}{is the minimum percent identity of the query relative to the
subject hits. The default is perID = 50.}

\item{reward}{is the reward for nucleotide match. The default is reward = 2.}

\item{num_threads}{is the number of CPUs to engage in the blastn search. The
default num_treads = NULL, uses \code{\link[parallel:detectCores]{parallel::detectCores()}} to determine the user's
number of CPUs automatically and use that for the value of -num_threads.}
}
\value{
a tibble 'output_table' representing the blastn results
}
\description{
Runs blastn with the input primer sequences converted to a primer_fasta file
as a query
}
\details{
Calls blastn with a primer fasta file as the query. The user can not add
additional search parameters, but can modify the available parameters.

run_primer_blastn takes a fasta file containing primers, uses blastn-short to
query them to a blast formatted database. The result is an output table with
the following columns of data: qseqid (query subject id), sgi (subject gi),
saccver (subject accession version), mismatch (number of mismatches between
the subject a query), sstart (subject start), send (subject end), staxids
(subject taxids).

Information about the blastn parameters can be found by accessing blastn -help
and at \href{https://www.ncbi.nlm.nih.gov/books/NBK279684/}{NCBI}.

Note:
The number of alignments returned for a given blast search is hardcoded at
"-num_alignments 10000000".
}
